Explicit casting to prevent accurcy loss (per warning)

+++ base/standard/src/main/java/org/openscience/cdk/tools/manipulator/AtomContainerSetManipulator.java
@@ -112,7 +112,7 @@ public class AtomContainerSetManipulator {
         for (int i = 0; i < set.getAtomContainerCount(); i++) {
             int thisCharge = AtomContainerManipulator.getTotalFormalCharge(set.getAtomContainer(i));
             double stoich = set.getMultiplier(i);
-            charge += stoich * thisCharge;
+            charge += stoich * (double)thisCharge;
         }
         return charge;
     }
@@ -126,7 +126,7 @@ public class AtomContainerSetManipulator {
         for (int i = 0; i < set.getAtomContainerCount(); i++) {
             int thisCharge = AtomContainerManipulator.getTotalFormalCharge(set.getAtomContainer(i));
             double stoich = set.getMultiplier(i);
-            charge += stoich * thisCharge;
+            charge += (int)(stoich * (double)thisCharge);
         }
         return charge;
     }
